Separated start, goal from traversables & now added functionality for style customisation
