+++ src/main_panda.py
@@ -11,8 +11,8 @@ if __name__ == "__main__":
             for z in range(0, 25):
                 map3d[x][y][z] = bool(random.getrandbits(1))
 
-    map3d[0][0][0] = True  # added for debugging purposes
-    map3d[0][0][1] = True  # added for debugging purposes
+    map3d[0][0][0] = False  # added for debugging purposes
+    map3d[0][0][1] = False  # added for debugging purposes
 
     view = MainView(map3d, start_pos=(0, 0, 0), goal_pos=(0, 0, 1))
